interface -> float to long

+++ Projecto/Interface/Interfaces.cs
@@ -13,7 +13,7 @@ namespace PADIDSTM
     {
         ServerHashTable requestHashTable();
         int addDataServer(string url);
-        float GetTimeStamp();
+        long GetTimeStamp();
 
     }
